Align test names with ccio.py move
