+++ test/io/testfilewriter.py
@@ -1,7 +1,7 @@
 # This file is part of cclib (http://cclib.github.io), a library for parsing
 # and interpreting the results of computational chemistry packages.
 #
-# Copyright (C) 2015, the cclib development team
+# Copyright (C) 2015-2016, the cclib development team
 #
 # The library is free software, distributed under the terms of
 # the GNU Lesser General Public version 2.1 or later. You should have
@@ -21,19 +21,16 @@ __filepath__ = os.path.realpath(__filedir__)
 __datadir__ = os.path.join(__filepath__, "..", "..")
 
 
-class XYZTest(unittest.TestCase):
-
-    def setUp(self):
-        self.XYZ = cclib.io.XYZ
+class WriterTest(unittest.TestCase):
 
     def test_init(self):
         """Does the class initialize correctly?"""
         fpath = os.path.join(__datadir__, "data/ADF/basicADF2007.01/dvb_gopt.adfout")
         data = cclib.io.ccopen(fpath).parse()
-        xyz = cclib.writer.xyzwriter.XYZ(data)
+        writer = cclib.io.filewriter.Writer(data)
 
         # The object should keep the ccData instance passed to its constructor.
-        self.assertEqual(xyz.ccdata, data)
+        self.assertEqual(writer.ccdata, data)
 
 
 if __name__ == "__main__":
