+++ test/test_io.py
@@ -1,7 +1,7 @@
 # This file is part of cclib (http://cclib.github.io), a library for parsing
 # and interpreting the results of computational chemistry packages.
 #
-# Copyright (C) 2015, the cclib development team
+# Copyright (C) 2015-2016, the cclib development team
 #
 # The library is free software, distributed under the terms of
 # the GNU Lesser General Public version 2.1 or later. You should have
@@ -13,7 +13,8 @@
 import sys
 import unittest
 
-sys.path.append('writer')
+sys.path.append('io')
+from testccio import *
 from testfilewriter import *
 from testxyzwriter import *
